Corrección de imports. La app ahora se ejecuta como un módulo.
