+++ app/Dockerfile
@@ -2,12 +2,12 @@ FROM facundogp/anom-api:v2-fastFM
 
 WORKDIR /code
 
-COPY ./requirements.txt ./requirements.txt
+ADD ./requirements.txt ./requirements.txt
 
 RUN pip install --no-cache-dir --upgrade -r ./requirements.txt
 
-COPY src ./app
+ADD src ./app/src
 
-WORKDIR /code/app
+WORKDIR /code/app/src
 
-CMD [ "python", "main.py"] 
+CMD ["python", "-m", "adetect"]
