+++ app/src/adetect/log.py
@@ -1,9 +1,10 @@
 import logging
 import os
 
+
 logging.basicConfig(level=logging.DEBUG,
                     format="%(asctime)s %(levelname)s %(name)s %(process)d: [%(threadName)s] %(message)s",
-                    filename='/tmp/anomalias.log',
+                    filename='anomalias.log',
                     filemode='w')
 
 logging.captureWarnings(True)
@@ -15,6 +16,7 @@ formatter = logging.Formatter('%(levelname)s: [%(threadName)s] %(message)s')
 console.setFormatter(formatter)
 
 logging.getLogger('').addHandler(console)
+
 
 def logger(s):
     return logging.getLogger(s)
