+++ utils/restart_services.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 docker stop anomalias_influxdb_1
 docker stop anomalias_api_1
-(cd ./../python_api/ && docker build -t service-api .)
+(cd ./../app/src/ && docker build -t service-api .)
 docker-compose up -d
